+++ packages/node/src/indexer/dictionary/v1/nearDictionaryV1.ts
@@ -1,4 +1,4 @@
-// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
+// Copyright 2020-2025 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
 import assert from 'assert';
@@ -13,7 +13,7 @@ import {
   NearCustomHandler,
   NearTransactionFilter,
 } from '@subql/common-near';
-import { NodeConfig, DictionaryV1 } from '@subql/node-core';
+import { NodeConfig, DictionaryV1, DsProcessorService } from '@subql/node-core';
 import {
   DictionaryQueryCondition,
   DictionaryQueryEntry as DictionaryV1QueryEntry,
@@ -26,7 +26,6 @@ import {
 import { sortBy, uniqBy } from 'lodash';
 import { SubqueryProject } from '../../../configure/SubqueryProject';
 import { isBaseHandler, isCustomHandler } from '../../../utils/project';
-import { DsProcessorService } from '../../ds-processor.service';
 
 type GetDsProcessor = DsProcessorService['getDsProcessor'];